Store the `precision` within the `WavefrontError`
Now it should be strictly typed as an Int of course. This will be useful
for when I implement wavefront error arithmetic.

+++ src/ScaleAperture.jl
@@ -40,14 +40,14 @@ function Π(v::Vector{T}, ε::T) where T <: Float64
 end
 
 function J(v::Vector{Float64}, ε::Float64;
-           precision = precision, finesse::Int = wavefront_finesse)
+           precision::Int = precision, finesse::Int = wavefront_finesse)
     v2, n_max = Π(v, ε)
     Zᵢ = similar(v, Polynomial)
     ΔW = Ψ(v2, Zᵢ, n_max; precision)
     Λ(ΔW; finesse)
 end
 
-function J(v::Vector{Float64}, ε::Float64, ::Type{Model}; precision = precision)
+function J(v::Vector{Float64}, ε::Float64, ::Type{Model}; precision::Int = precision)
     v2, n_max = Π(v, ε)
     Zᵢ = similar(v, Polynomial)
     Ψ(v2, Zᵢ, n_max; precision)
